refactor: make it make sense, and make it pretty!
Tests don't work yet, though! :D

refs: BLAIS5-4631
